// repositorio/SIMTR/4-implementacao/simtr-web/src/app/documento/enum-documento/situacao-documento.enum.ts
export enum SituacaoDocumento {
    CONFORME = "Conforme",
    CRIADO = "Criado",
    IGNORADO = "Ignorado",
    INVALIDO = "Invalido",
    PENDENTE = "Pendente",
    QUARENTENA = "Quarentena",
    RECUSADO = "Recusado",
    REJEITADO = "Rejeitado",
    REPROVADO = "Reprovado",
    SUBSTITUIDO = "Substituido"
}